+++ packages/mcp-server/src/permissions.parseScopes.test.ts
@@ -1,5 +1,5 @@
 import { describe, it, expect } from "vitest";
-import { parseScopes, type Scope } from "./permissions";
+import { parseScopes, expandScopes, type Scope } from "./permissions";
 
 describe("parseScopes", () => {
   it("parses comma-separated string with trimming and de-dup", () => {
@@ -16,7 +16,7 @@ describe("parseScopes", () => {
     const { valid, invalid } = parseScopes([
       "member:read",
       "x",
-      123 as any,
+      123 as unknown,
       " team:write ",
       "",
     ]);
@@ -33,5 +33,31 @@ describe("parseScopes", () => {
       invalid: [],
     });
     expect(parseScopes([])).toEqual({ valid: new Set<Scope>(), invalid: [] });
+  });
+});
+
+// Consolidated strict-like parseScopes cases
+describe("parseScopes (strict-like cases)", () => {
+  it("returns invalid tokens for unknown scopes", () => {
+    const { valid, invalid } = parseScopes("foo,bar,org:admin");
+    expect(invalid).toEqual(["foo", "bar"]);
+    expect([...valid]).toContain("org:admin");
+  });
+
+  it("returns only valid set when all are valid", () => {
+    const { valid, invalid } = parseScopes("event:admin,org:read");
+    expect(invalid).toEqual([]);
+    const out = new Set<Scope>(valid);
+    expect(out.has("event:admin")).toBe(true);
+    expect(out.has("org:read")).toBe(true);
+  });
+});
+
+// Related behavior validation for expandScopes
+describe("expandScopes", () => {
+  it("includes implied lower scopes", () => {
+    const expanded = expandScopes(new Set<Scope>(["event:write"]));
+    expect(expanded.has("event:read")).toBe(true);
+    expect(expanded.has("event:write")).toBe(true);
   });
 });
